requirements.txt
streamlite==1.36.0
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
numpy==2.0.0
prophet==1.1.5